Add "Show Map" GUI option
This commit adds a new GUI option to toggle between displaying only
the points of the current frame or the entire map.
Enabling this option makes it possible to visualize the overall map and
can be useful in debugging and analysis. The new option can be accessed
through the GUI controls and is off by default.